docs(change): from reset --hard to update recursive
This command initializes and updates submodules. If a submodule is not initialized, it initializes it. If a submodule is already initialized, it updates it to the commit specified by the superproject. It is often used after cloning a repository with submodules or after changes to the submodule references in the superproject.
Moving away from using `git reset --hard`, because it can be a destructive operation as it discards local changes.